Setup visual studio project

+++ src/util.c
@@ -1,9 +1,9 @@
 
 static inline void DrawTextureCenteredV(Texture2D texture, Vector2 position, Color tint) {
-	DrawTextureV(texture, Vector2Subtract(position, (Vector2) { texture.width >> 1, texture.height >> 1 }), tint);
+	DrawTextureV(texture, Vector2Subtract(position, (Vector2) { (f32)(texture.width >> 1), (f32)(texture.height >> 1) }), tint);
 }
 
 static inline void DrawTextDefault(const char* text, Vector2 position, Color tint) {
 	const Font font = AssetsGetFont();
-	DrawTextEx(font, text, position, font.baseSize, 0.0f, tint);
+	DrawTextEx(font, text, position, (f32)font.baseSize, 0.0f, tint);
 }
